==> notebooks/01_md_scraper.ipynb <==
### **NWS SPC Mesoscale Discussion Scraper**
#### **Created by Tyson Stewart for use in EAE 598/698**
##### Data courtesy of Iowa Environmental Mesonet

Import necessary packages and read CSV of MDs and their relevant details.

import geopandas as gp
import pandas as pd
import urllib.request

csv = pd.read_csv("central_ok_mds.csv")

#Create new column called "Watch Parameter" that by default will be set to all zeroes unless a watch is present
csv["WATCH PARAMETER"] = [0] * len(csv["CONCERN"])

csv

MD CSV does not contain the actual text, and is inconsistent regarding watch probability and MD type. To fix this we can use IEM's API to scrape each MD for the relevant details.

# Lists to have our scrapped data inserted into
concern_list = []
prob_list = []

#List of terms in the "Concerning" line that indicate that the MD is NOT a watch
watch_tag = ["Unlikely", "Likely", "Possible", "UNLIKELY", "LIKELY", "POSSIBLE", "unlikely", "likely", "possible", "needed soon", "NEEDED SOON"]

c = 0

for i in range(len(csv["SPC PRODUCT ID"])):
    #Pull the MD text from IEM API
    prod_id = csv["SPC PRODUCT ID"][i]
    link = f"https://mesonet.agron.iastate.edu/api/1/nwstext/{prod_id}"
    text = urllib.request.urlopen(link)
    #Create/reset flag variables for Watch/MD processing
    c_flag = 0
    w_flag = 0
    c += 1
    
    #Check every line in the MD Text to find "concerning" and watch probability lines
    for line in text:
        if ("CONCERNING" in str(line) or "Concerning" in str(line)) and c_flag == 0:
            d = 0
            concern = str(line[13:])
            concern_strip = concern.strip("\n")[2:-4]

            concern_list.append(concern_strip)
            #print(f"{prod_id} is {concern_strip}")

            if ("WATCH" in concern_strip) or ("Watch" in concern_strip):
                verify_flag = 1
                for i in watch_tag:
                    #Check if any of our watch disqualifying terms (watch_tag) are in the concerning tag.
                    if watch_tag[d] in concern_strip:
                        verify_flag = 0
                        break
                    d += 1
                
                # If "Watch" is in the concerning tag, and there is none of the watch disqualifying terms are present, it must be an active watch.
                if verify_flag == 1:
                    w_flag = 1
                    prob_list.append(100)
                    #print("ACTIVE WATCH!")

            c_flag = 1
        
        if "PROBABILITY OF WATCH ISSUANCE" in str(line) or "Probability of Watch Issuance" in str(line):
            watch_prob = str(line)[34:-11]
            #print(f"Watch prob is {watch_prob}%")
            prob_list.append(watch_prob)
            w_flag = 1
    
    #If no watch probability data is present, assume that it is 0%.
    if w_flag == 0:
        watch_prob = 0
        #print(f"Watch prob is {watch_prob}%")
        prob_list.append(watch_prob)

#concern_list

***Unit Testing:*** Verify that the watch probability and MD type lists match in length.

print(len(concern_list))
print(len(prob_list))

Iterates through each list and applies the value into the CSV. Also check if the watch probability is 100% and if so set the "Watch Parameter" value to 1. (Watch Parameter will serve as our binary classifier for the ML model).

for i in range(len(prob_list)):
    csv.loc[i, "CONFIDEN"] = prob_list[i]

    if csv["CONFIDEN"][i] == 100:
        csv.loc[i, "WATCH PARAMETER"] = 1

for i in range(len(prob_list)):
    csv.loc[i, "CONCERN"] = concern_list[i]
csv

Filter the newly modified CSV file to only contain MDs concerning severe thunderstorm and tornado threats.

sub_csv = csv[(csv["CONCERN"].str.contains("SEVERE")) | (csv["CONCERN"].str.contains("Severe")) | (csv["CONCERN"].str.contains("TORNADO")) | (csv["CONCERN"].str.contains("Tornado"))]

sub_csv

***Unit Testing:*** Evaluate CSV fields to ensure that there are no null data points.

sub_csv.reset_index(drop=True, inplace=True)

key_list = list(sub_csv.keys())

# Flag variable to raise if a null value is detected
null_flag = 0

#Test each key to ensure all data is present
for i in key_list:
    for j in range(0, (len(sub_csv[i]) - 1)):
        if sub_csv.loc[j, i] == "NaN" or sub_csv.loc[j, i] == "Null" or sub_csv.loc[j, i] == None:
            null_flag = 1
            print(f"NULL VALUE DETECTED AT INDEX {j}!")

if null_flag == 0:
    print("No null values detected. Check passed.")

Write the newly filtered CSV to disk.

sub_csv.to_csv("central_ok_mds_FINAL.csv")
