==> notebooks/13_random_forest_model_metrics .ipynb <==
# This notebook goes through process of training Random Forest model and displays model metrics

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
import joblib
import sys

# Unit test to determine if file has all data necessary to run Random Forest model

def csv_test(df):
    '''Test to ensure all environmental data is in csv file and none is missing'''
    if df.isnull().sum().all() == 0:
        print('✅ Data is correctly read in and ready for training!')
    else:
        sys.exit('❌ Error. One or more values are missing. Please check csv file to remove any dates with incomplete environmental values.')

# Read in environmental and MD data

# Read in dataframe
df = pd.read_csv("central_ok_mds_env_FINAL.csv")

# Complete unit test
csv_test(df)

# Split data into training, validation, and testing

# Split data
train_data, val_test_data = train_test_split(df, test_size = 0.3, random_state = 988)
val_data, test_data = train_test_split(val_test_data, test_size = 0.6667, random_state = 988)

# Train base Random Forest model

# Set feature list of MD and RAP analysis data
feature_list = ['month', 'time', 'pwat', 'mslp', 'cape', 'cin', 't2m', 'rh2m', 'q2m', 'u10', 'v10', 'uv10', 'lcl', 'shr0_6', 'srh0_1', 'srh0_3']

# Train base Random Forest model
rf = RandomForestClassifier(random_state = 56)
rf.fit(train_data[feature_list].values, train_data.label.values)

# Test (using validation data) base Random Forest model
predicted = rf.predict(val_data[feature_list].values)
expected = val_data.label.values

# Print classification report
print('Classification report for classifier %s:\n%s\n'
      % (rf, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(val_data[feature_list].values.shape[1]):
    print('%d. (%s) feature %d (%f)' % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf, 'random_forest_base.joblib')

# Optimized Random Forest model

# Create parameterizations to test on training data for optimization
param_grid = [{
    'n_estimators': [20, 40, 60, 80, 100, 150, 200],
    'criterion': ['gini', 'entropy', 'log_loss'],
    'min_samples_split': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
    'min_samples_leaf': [1, 2, 3, 4, 5, 6, 7, 8],
    'max_depth': [2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 14, 16, 18, 20] }]

# Fit optizimation to training data
grid_search = GridSearchCV(rf, param_grid, cv = 2, scoring = 'accuracy', n_jobs = -1)

grid_search.fit(train_data[feature_list].values, train_data.label.values)
grid_search.best_params_

# Train optimized Random Forest model
rf_opt = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 700)
rf_opt.fit(train_data[feature_list].values, train_data.label.values)

# Test (using validation data) optimized Random Forest model
predicted = rf_opt.predict(val_data[feature_list].values)
expected = val_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_opt, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_opt.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(val_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file
# joblib.dump(rf_opt, "random_forest_opt.joblib")

# Testing and perturbing opitmized Random Forest model

# Train perturbed final Random Forest model
rf_final0 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 136)
rf_final0.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final0.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final0, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

importances = rf_final0.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))

# Save Random Forest model as joblib file
# joblib.dump(rf_final0, "random_forest_final0.joblib")

# Train perturbed final Random Forest model
rf_final1 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 2)
rf_final1.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final1.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final1, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final1.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))

# Save Random Forest model as joblib file
# joblib.dump(rf_final1, "random_forest_final1.joblib")

# Train perturbed final Random Forest model
rf_final2 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 439)
rf_final2.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final2.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final2, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final2.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final2, "random_forest_final2.joblib")

# Train perturbed final Random Forest model
rf_final3 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 712)
rf_final3.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final3.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final3, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final3.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final3, "random_forest_final3.joblib")

# Train perturbed final Random Forest model
rf_final4 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 297)
rf_final4.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final4.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final4, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final4.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final4, "random_forest_final4.joblib")

# Train perturbed final Random Forest model
rf_final5 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 331)
rf_final5.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final5.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final5, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final5.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final5, "random_forest_final5.joblib")

# Train perturbed final Random Forest model
rf_final6 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 984)
rf_final6.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final6.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final6, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final6.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final6, "random_forest_final6.joblib")

# Train perturbed final Random Forest model
rf_final7 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 813)
rf_final7.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final7.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final7, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final7.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final7, "random_forest_final7.joblib")

# Train perturbed final Random Forest model
rf_final8 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 47)
rf_final8.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final8.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final8, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final8.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final8, "random_forest_final8.joblib")

# Train perturbed final Random Forest model
rf_final9 = RandomForestClassifier(criterion = 'log_loss', max_depth = 16, min_samples_leaf = 6, min_samples_split = 4, n_estimators = 40, random_state = 454)
rf_final9.fit(train_data[feature_list].values, train_data.label.values)

# Test perturbed final Random Forest model
predicted = rf_final9.predict(test_data[feature_list].values)
expected = test_data.label.values

# Print classification report
print("Classification report for classifier %s:\n%s\n"
      % (rf_final9, metrics.classification_report(expected, predicted)))

# Print brier score
print(f'Brier score: {metrics.brier_score_loss(expected, predicted)}', end = '\n\n')

# Print feature importance precentages
importances = rf_final9.feature_importances_

indices = np.argsort(importances)[::-1]

for i in range(test_data[feature_list].values.shape[1]):
    print("%d. (%s) feature %d (%f)" % (i+1, feature_list[indices[i]], indices[i], importances[indices[i]]))
    
# Save Random Forest model as joblib file    
# joblib.dump(rf_final9, "random_forest_final9.joblib")
